Adapt to changes to ledger-fetch, json-types and hooks
Had an issue building with the new optional chaining syntax in latest
TS. For now I reverted the uses of optional chaining, but I suspect it's
a failure of the project setup that this does not work. I tried updating
to latest TS version (reflected in package.json and yarn.lock) but that
did not seem to work.